// src/types/artifact.ts
export type Artifact = {
    id: string
    name: string
    key: string
    rarity_max: number
    rarity_min: number
    release_version: string
    flower_description?: string
    feather_description?: string
    sand_description?: string
    goblet_description?: string
    circlet_description?: string
    flower_name?: string
    feather_name?: string
    sand_name?: string
    goblet_name?: string
    circlet_name?: string
    two_pc_bonus?: string
    four_pc_bonus?: string
}